# docs/dev3/weather-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "HackAlem Weather Service",
    "description": "\nАрхивные прогнозы погоды по точке ВЭС «Шелек» в том виде, в каком они были доступны на момент\nпрогноза, и почасовая история турбин (SCADA).\n\n**Кто вызывает.** Backend dev1 по `WEATHER_SERVICE_URL`. Строки `GET /nwp` уходят в `POST /predict`\nML-сервиса как есть. Frontend к сервису напрямую не ходит. Общая схема: `docs/api-contract.md`.\n\n**Главное правило.** Ни одна строка погоды не опубликована позже `as_of`: прогон считается доступным\nв момент `run_init_utc + задержка публикации источника`. Перед отдачей ответа это проверяется еще раз.\n\n**Время.** ISO 8601 в UTC с `Z`: `2026-01-31T02:00:00Z`. Время без пояса дает 422, время с другим\nсмещением переводится в UTC. Пустое значение — `null`.\n\n**Источники.** `ifs`, `ifs025`, `gfs`, `icon`, `gem` и `ensemble` — среднее `ifs025`, `gfs`, `icon`, `gem`.\nПодробности в `GET /sources`.\n\n**Ошибки.** Всегда конверт `{\"error\": {\"code\", \"message\", \"details\"}}`, как в backend.\n",
    "version": "1.0.0"
  },
  "paths": {
...
    "/nwp": {
      "get": {
        "tags": [
          "Weather"
        ],
        "summary": "Погода, опубликованная к моменту as_of",
        "description": "Для каждого часа окна — самый свежий прогон с `available_at_utc <= as_of`. По умолчанию окно — горизонт выпуска `as_of + 1 ч … as_of + 48 ч`, ровно то, что ждет `POST /predict`. `from` и `to` включительно, начало часа, не больше 168 ч.\n\nОдин источник без прогона дает 404 `NO_RUN_AVAILABLE`, частичной таблицы не бывает. Из нескольких источников пропускаются те, у которых прогона нет, они перечислены в `missing`.\n\n`ensemble` — среднее `ifs025`, `gfs`, `icon`, `gem`: `ws80` — ветер на высоте ступицы (80 м, без него 100 или 120 м с показателем сдвига 0,14), `ws_spread` — разброс между моделями. **Строки `ensemble` не отправляйте в `/predict` вместе со строками участников:** ML-сервис сам усредняет источники, и ансамбль учтется дважды.\n\n`lead_h` в строке — часы от запуска прогона, а не от выпуска.",
        "operationId": "nwp_nwp_get",
        "parameters": [
          {
            "name": "source",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Один или несколько источников через запятую: ifs, ifs025, gfs, icon, gem, ensemble",
              "examples": [
                "gfs,ensemble"
              ],
              "title": "Source"
            },
            "description": "Один или несколько источников через запятую: ifs, ifs025, gfs, icon, gem, ensemble"
          },
          {
            "name": "as_of",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time",
              "description": "Момент прогноза: только прогоны, опубликованные не позже него",
              "examples": [
                "2026-01-31T02:00:00Z"
              ],
              "title": "As Of"
            },
            "description": "Момент прогноза: только прогоны, опубликованные не позже него"
          },
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Первый час окна. По умолчанию as_of + 1 ч",
              "title": "From"
            },
            "description": "Первый час окна. По умолчанию as_of + 1 ч"
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Последний час окна. По умолчанию from + 47 ч",
              "title": "To"
            },
            "description": "Последний час окна. По умолчанию from + 47 ч"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NwpResponse"
                },
                "example": {
                  "as_of_utc": "2026-01-31T02:00:00Z",
                  "from_utc": "2026-01-31T03:00:00Z",
                  "to_utc": "2026-02-02T02:00:00Z",
                  "sources": [
                    "gfs",
                    "ensemble"
                  ],
                  "missing": [],
                  "rows": [
                    {
                      "valid_time_utc": "2026-01-31T03:00:00Z",
                      "source": "gfs",
                      "run_init_utc": "2026-01-30T00:00:00Z",
                      "available_at_utc": "2026-01-30T07:00:00Z",
                      "lead_h": 27,
                      "ws80": 13.87,
                      "ws100": 14.63,
                      "ws120": 14.93,
                      "wd100": 65.0,
                      "gust10": 17.2,
                      "t2m": -2.3,
                      "rh2m": 49.0,
                      "psfc": 956.1
                    },
                    {
                      "valid_time_utc": "2026-01-31T03:00:00Z",
                      "source": "ensemble",
                      "run_init_utc": "2026-01-30T00:00:00Z",
                      "available_at_utc": "2026-01-30T10:00:00Z",
                      "lead_h": 27,
                      "ws80": 10.75,
                      "wd100": 68.0,
                      "gust10": 10.17,
                      "t2m": -3.8,
                      "rh2m": 62.25,
                      "psfc": 954.62,
                      "ws_spread": 2.83,
                      "members": [
                        "gem",
                        "gfs",
                        "icon",
                        "ifs025"
                      ]
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Нет прогона, опубликованного к as_of",
            "content": {
              "application/json": {
                "examples": {
                  "no_run": {
                    "summary": "Нет прогона",
                    "value": {
                      "error": {
                        "code": "NO_RUN_AVAILABLE",
                        "message": "На момент as_of нет опубликованного прогона хотя бы для одного из запрошенных часов",
                        "details": {
                          "source": "ifs",
                          "as_of_utc": "2024-03-15T02:00:00Z",
                          "missing_hours": 48,
                          "examples": [
                            "2024-03-15T03:00:00Z"
                          ]
                        }
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Неизвестный источник или ошибка в параметрах",
            "content": {
              "application/json": {
                "examples": {
                  "unknown": {
                    "summary": "Неизвестный источник",
                    "value": {
                      "error": {
                        "code": "UNKNOWN_SOURCE",
                        "message": "Неизвестный источник погоды: ecmwf_ifs",
                        "details": {
                          "unknown": [
                            "ecmwf_ifs"
                          ],
                          "known": [
                            "ifs",
                            "ifs025",
                            "gfs",
                            "icon",
                            "gem",
                            "ensemble"
                          ]
                        }
                      }
                    }
                  },
                  "validation": {
                    "summary": "Ошибка в параметрах",
                    "value": {
                      "error": {
                        "code": "VALIDATION_ERROR",
                        "message": "Окно больше 168 ч",
                        "details": {
                          "from": "2026-01-31T03:00:00Z",
                          "to": "2026-02-10T03:00:00Z"
                        }
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Сработала проверка утечки будущего или внутренняя ошибка",
            "content": {
              "application/json": {
                "examples": {
                  "leakage": {
                    "summary": "Проверка утечки",
                    "value": {
                      "error": {
                        "code": "LEAKAGE_GUARD",
                        "message": "Ответ остановлен последней проверкой на утечку будущего: …",
                        "details": {}
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/scada": {
      "get": {
        "tags": [
          "SCADA"
        ],
        "summary": "Почасовая история турбин",
        "description": "Интервал `[from, until)`, не больше 31 суток. Время уже переведено в UTC из UTC+6 файлов SCADA. Данные заканчиваются 31.01.2026 17:00 UTC, для более поздних часов придет пустой список.",
        "operationId": "scada_scada_get",
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time",
              "description": "Начало интервала, включительно",
              "examples": [
                "2026-01-01T00:00:00Z"
              ],
              "title": "From"
            },
            "description": "Начало интервала, включительно"
          },
          {
            "name": "until",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time",
              "description": "Конец интервала, не включительно",
              "examples": [
                "2026-02-01T00:00:00Z"
              ],
              "title": "Until"
            },
            "description": "Конец интервала, не включительно"
          },
          {
            "name": "turbine",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "T1",
                    "T2"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "T1 или T2. По умолчанию обе",
              "title": "Turbine"
            },
            "description": "T1 или T2. По умолчанию обе"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ScadaRow"
                  },
                  "title": "Response Scada Scada Get"
                },
                "example": [
                  {
                    "time_utc": "2026-01-30T20:00:00Z",
                    "turbine": "T1",
                    "power_norm": 0.1867,
                    "wind_ms": 5.73,
                    "temp_c": -1.54,
                    "flag": ""
                  },
                  {
                    "time_utc": "2026-01-28T09:00:00Z",
                    "turbine": "T1",
                    "power_norm": 0.01,
                    "wind_ms": 4.43,
                    "temp_c": 0.05,
                    "flag": "icing"
                  }
                ]
              }
            }
          },
          "422": {
            "description": "Неизвестный источник или ошибка в параметрах",
            "content": {
              "application/json": {
                "examples": {
                  "validation": {
                    "summary": "Ошибка в параметрах",
                    "value": {
                      "error": {
                        "code": "VALIDATION_ERROR",
                        "message": "Окно больше 168 ч",
                        "details": {
                          "from": "2026-01-31T03:00:00Z",
                          "to": "2026-02-10T03:00:00Z"
                        }
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Данные не загружены",
            "content": {
              "application/json": {
                "examples": {
                  "no_scada": {
                    "summary": "SCADA не загружена",
                    "value": {
                      "error": {
                        "code": "DATA_UNAVAILABLE",
                        "message": "SCADA не загружена: нет файлов турбин в DATA_DIR",
                        "details": {}
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ErrorBody": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "description": "NO_RUN_AVAILABLE, UNKNOWN_SOURCE, VALIDATION_ERROR, LEAKAGE_GUARD, DATA_UNAVAILABLE, INTERNAL_ERROR"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "additionalProperties": true,
            "type": "object",
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse"
      },
...
      "MissingSource": {
        "properties": {
          "source": {
            "type": "string",
            "title": "Source"
          },
          "missing_hours": {
            "type": "integer",
            "title": "Missing Hours",
            "description": "Сколько часов окна у источника нет"
          }
        },
        "type": "object",
        "required": [
          "source",
          "missing_hours"
        ],
        "title": "MissingSource"
      },
      "NwpResponse": {
        "properties": {
          "as_of_utc": {
            "type": "string",
            "format": "date-time",
            "title": "As Of Utc"
          },
          "from_utc": {
            "type": "string",
            "format": "date-time",
            "title": "From Utc"
          },
          "to_utc": {
            "type": "string",
            "format": "date-time",
            "title": "To Utc"
          },
          "sources": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Sources",
            "description": "Источники, которые попали в ответ"
          },
          "missing": {
            "items": {
              "$ref": "#/components/schemas/MissingSource"
            },
            "type": "array",
            "title": "Missing",
            "description": "Запрошенные источники, пропущенные из-за отсутствия прогона"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/NwpRow"
            },
            "type": "array",
            "title": "Rows",
            "description": "По строке на час и источник, по возрастанию valid_time_utc"
          }
        },
        "type": "object",
        "required": [
          "as_of_utc",
          "from_utc",
          "to_utc",
          "sources",
          "missing",
          "rows"
        ],
        "title": "NwpResponse"
      },
...
      "ScadaRow": {
        "properties": {
          "time_utc": {
            "type": "string",
            "format": "date-time",
            "title": "Time Utc",
            "description": "Начало часа, UTC. Время SCADA уже переведено из UTC+6"
          },
          "turbine": {
            "type": "string",
            "enum": [
              "T1",
              "T2"
            ],
            "title": "Turbine"
          },
          "power_norm": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Power Norm",
            "description": "Средняя мощность за час, доля номинала 0…1"
          },
          "wind_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wind Ms",
            "description": "Средний ветер анемометра турбины, м/с"
          },
          "temp_c": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Temp C",
            "description": "Средняя температура датчика турбины, °C"
          },
          "flag": {
            "type": "string",
            "title": "Flag",
            "description": "\"\" — чистый час; downtime, curtailment, icing, frozen_sensor, cut_out"
          }
        },
        "type": "object",
        "required": [
          "time_utc",
          "turbine",
          "power_norm",
          "wind_ms",
          "temp_c",
          "flag"
        ],
        "title": "ScadaRow"
      },
...
    }
  }
}